# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FPS Brandpreventie - Platform voor brandpreventieve gebouwvoorzieningen
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gebouwen/{id}/emails:
    get:
      operationId: listGebouwEmails
      tags: [emails]
      summary: E-mailarchief van een gebouw
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GebouwEmail"
        "403":
          description: Geen toegang
    post:
      operationId: createGebouwEmail
      tags: [emails]
      summary: E-mail uploaden en verwerken (beheerder+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GebouwEmailInput"
      responses:
        "201":
          description: Aangemaakt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GebouwEmail"
        "403":
          description: Geen toegang
  /gebouwen/{id}/emails/samenvatting:
    get:
      operationId: getGebouwEmailSamenvatting
      tags: [emails]
      summary: Gecombineerde AI-projectsamenvatting van alle e-mails
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GebouwEmailSamenvatting"
        "404":
          description: Nog geen samenvatting beschikbaar
        "403":
          description: Geen toegang
    post:
      operationId: generateGebouwEmailSamenvatting
      tags: [emails]
      summary: AI-projectsamenvatting genereren of bijwerken (beheerder+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GebouwEmailSamenvatting"
        "403":
          description: Geen toegang
    patch:
      operationId: updateGebouwEmailSamenvatting
      tags: [emails]
      summary: Projectsamenvatting bewerken en bevestigen (beheerder+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GebouwEmailSamenvattingUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GebouwEmailSamenvatting"
        "403":
          description: Geen toegang
  /gebouwen/{id}/emails/{emailId}:
    get:
      operationId: getGebouwEmail
      tags: [emails]
      summary: E-mail ophalen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: emailId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GebouwEmail"
        "404":
          description: Niet gevonden
    delete:
      operationId: deleteGebouwEmail
      tags: [emails]
      summary: E-mail verwijderen (beheerder+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: emailId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Verwijderd
components:
  schemas:
...
    GebouwEmailBijlage:
      type: object
      required: [id, email_id, bestandsnaam]
      properties:
        id:
          type: integer
        email_id:
          type: integer
        bestandsnaam:
          type: string
        object_pad:
          type: ["string", "null"]
        content_type:
          type: ["string", "null"]
        grootte:
          type: ["integer", "null"]
    GebouwEmail:
      type: object
      required: [id, gebouw_id, bestandsnaam, status, aangemaakt_op]
      properties:
        id:
          type: integer
        gebouw_id:
          type: integer
        bestandsnaam:
          type: string
        object_pad:
          type: ["string", "null"]
        afzender:
          type: ["string", "null"]
        ontvanger:
          type: ["string", "null"]
        onderwerp:
          type: ["string", "null"]
        datum:
          type: ["string", "null"]
        inhoud_tekst:
          type: ["string", "null"]
        ai_omschrijving:
          type: ["string", "null"]
        ai_naw:
          type: ["string", "null"]
        ai_contactinfo:
          type: ["string", "null"]
        ai_tekeningen:
          type: ["string", "null"]
        ai_actiepunten:
          type: ["string", "null"]
        status:
          type: string
        aangemaakt_op:
          type: string
        bijlagen:
          type: array
          items:
            $ref: "#/components/schemas/GebouwEmailBijlage"
    GebouwEmailInput:
      type: object
      required: [object_pad, bestandsnaam]
      properties:
        object_pad:
          type: string
        bestandsnaam:
          type: string
    GebouwEmailSamenvatting:
      type: object
      required: [id, gebouw_id, aantal_emails, geverifieerd, bijgewerkt_op]
      properties:
        id:
          type: integer
        gebouw_id:
          type: integer
        opdrachtomschrijving:
          type: ["string", "null"]
        opdrachtgever:
          type: ["string", "null"]
        contactgegevens:
          type: ["string", "null"]
        afspraken:
          type: ["string", "null"]
        actiepunten:
          type: ["string", "null"]
        besluiten:
          type: ["string", "null"]
        tekeningen:
          type: ["string", "null"]
        risicos:
          type: ["string", "null"]
        contactpersonen:
          type: array
          items:
            $ref: "#/components/schemas/EmailContactpersoon"
        aantal_emails:
          type: integer
        geverifieerd:
          type: boolean
        gecontroleerd_door:
          type: ["string", "null"]
        gecontroleerd_op:
          type: ["string", "null"]
        bijgewerkt_op:
          type: string
    GebouwEmailSamenvattingUpdate:
      type: object
      properties:
        opdrachtomschrijving:
          type: ["string", "null"]
        opdrachtgever:
          type: ["string", "null"]
        contactgegevens:
          type: ["string", "null"]
        afspraken:
          type: ["string", "null"]
        actiepunten:
          type: ["string", "null"]
        besluiten:
          type: ["string", "null"]
        tekeningen:
          type: ["string", "null"]
        risicos:
          type: ["string", "null"]
        geverifieerd:
          type: boolean
        contactpersonen:
          type: array
          items:
            $ref: "#/components/schemas/EmailContactpersoon"
    EmailContactpersoon:
      type: object
      required: [rol, naam]
      properties:
        rol:
          type: string
          description: opdrachtgever, gebruiker, installateur, aannemer, eigenaar of aanvrager
        naam:
          type: string
        organisatie:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        telefoon:
          type: ["string", "null"]
        functie:
          type: ["string", "null"]
          description: Functietitel binnen de organisatie
        status:
          type: ["string", "null"]
          description: voorstel, bevestigd of afgewezen
        relevantie:
          type: ["string", "null"]
          description: relevant of ter_controle
        bron_email_id:
          type: ["integer", "null"]
          description: ID van de e-mail waaruit dit contact is geextraheerd
        bron_onderwerp:
          type: ["string", "null"]
          description: Onderwerp van de bronemail

    # ── DASHBOARD ─────────────────────────────────────────────────────────
...